Affiliate-Bot/requirements.txt
anyio==3.6.2
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
blinker==1.5
Brotli==1.0.9
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
colorama==0.4.6
cryptography==39.0.0
cssselect==1.2.0
dnspython==2.2.1
et-xmlfile==1.1.0
exceptiongroup==1.0.1
geographiclib==2.0
geopy==2.3.0
gnureadline==8.1.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.2
httpx==0.23.1
hyperframe==6.0.1
idna==2.10
instagram-private-api==1.6.0.0
kaitaistruct==0.10
lxml==4.9.2
numpy==1.24.1
openpyxl==3.0.10
outcome==1.2.0
pandas==1.5.2
prettytable==0.7.2
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.6.0
pymongo==4.3.3
pyOpenSSL==23.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.24.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rsa==4.9
selenium==4.6.0
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
Telethon==1.26.1
trio==0.22.0
trio-websocket==0.9.2
twilio==7.16.2
undetected-chromedriver==3.2.1
urllib3==1.25.11
websockets==10.4
wsproto==1.2.0
zstandard==0.19.0
